refactor: position page initial loading
